Function to manipulate print setup.

Other settings are available but not exposed. Please see the java docs.

printSetup(
  sheet,
  fitHeight = NULL,
  fitWidth = NULL,
  copies = NULL,
  draft = NULL,
  footerMargin = NULL,
  headerMargin = NULL,
  landscape = FALSE,
  pageStart = NULL,
  paperSize = NULL,
  noColor = NULL
)

Arguments

sheet

a worksheet object Worksheet.

fitHeight

numeric value to set the number of pages high to fit the sheet in.

fitWidth

numeric value to set the number of pages wide to fit the sheet in.

copies

numeric value to set the number of copies.

draft

logical indicating if it's a draft or not.

footerMargin

numeric value to set the footer margin.

headerMargin

numeric value to set the header margin.

landscape

logical value to specify the paper orientation.

pageStart

numeric value from where to start the page numbering.

paperSize

character to set the paper size. Valid values are "A4_PAPERSIZE", "A5_PAPERSIZE", "ENVELOPE_10_PAPERSIZE", "ENVELOPE_CS_PAPERSIZE", "ENVELOPE_DL_PAPERSIZE", "ENVELOPE_MONARCH_PAPERSIZE", "EXECUTIVE_PAPERSIZE", "LEGAL_PAPERSIZE", "LETTER_PAPERSIZE".

noColor

logical value to indicate if the prints should be color or not.

Value

A reference to a java PrintSetup object.

Author

Adrian Dragulescu

Examples

wb <- createWorkbook()
sheet <- createSheet(wb, "Sheet1")
ps   <- printSetup(sheet, landscape=TRUE, copies=3)