Functions to manipulate rows of a worksheet.

removeRow is just a convenience wrapper to remove the rows from the sheet (before saving). Internally it calls lapply.

createRow(sheet, rowIndex = 1:5)

getRows(sheet, rowIndex = NULL)

removeRow(sheet, rows = NULL)

setRowHeight(rows, inPoints, multiplier = NULL)

Arguments

sheet

a worksheet object as returned by createSheet or by subsetting getSheets.

rowIndex

a numeric vector specifying the index of rows to create. For getRows, a NULL value will return all non empty rows.

rows

a list of Row objects.

inPoints

a numeric value to specify the height of the row in points.

multiplier

a numeric value to specify the multiple of default row height in points. If this value is set, it takes precedence over the inPoints argument.

Value

For getRows a list of java object references each pointing to a row. The list is named with the row number.

See also

Author

Adrian Dragulescu

Examples

file <- system.file("tests", "test_import.xlsx", package = "xlsx")

wb <- loadWorkbook(file)
sheets <- getSheets(wb)

sheet <- sheets[[2]]
rows  <- getRows(sheet)  # get all the rows

# see all the available java methods that you can call
rJava::.jmethods(rows[[1]])
#>  [1] "public int org.apache.poi.xssf.usermodel.XSSFRow.getPhysicalNumberOfCells()"                                                                       
#>  [2] "public org.apache.poi.ss.usermodel.CellStyle org.apache.poi.xssf.usermodel.XSSFRow.getRowStyle()"                                                  
#>  [3] "public org.apache.poi.xssf.usermodel.XSSFCellStyle org.apache.poi.xssf.usermodel.XSSFRow.getRowStyle()"                                            
#>  [4] "public void org.apache.poi.xssf.usermodel.XSSFRow.setRowStyle(org.apache.poi.ss.usermodel.CellStyle)"                                              
#>  [5] "public java.util.Iterator org.apache.poi.xssf.usermodel.XSSFRow.cellIterator()"                                                                    
#>  [6] "public org.apache.poi.xssf.usermodel.XSSFCell org.apache.poi.xssf.usermodel.XSSFRow.getCell(int,org.apache.poi.ss.usermodel.Row$MissingCellPolicy)"
#>  [7] "public org.apache.poi.ss.usermodel.Cell org.apache.poi.xssf.usermodel.XSSFRow.getCell(int,org.apache.poi.ss.usermodel.Row$MissingCellPolicy)"      
#>  [8] "public org.apache.poi.xssf.usermodel.XSSFCell org.apache.poi.xssf.usermodel.XSSFRow.getCell(int)"                                                  
#>  [9] "public org.apache.poi.ss.usermodel.Cell org.apache.poi.xssf.usermodel.XSSFRow.getCell(int)"                                                        
#> [10] "public void org.apache.poi.xssf.usermodel.XSSFRow.removeCell(org.apache.poi.ss.usermodel.Cell)"                                                    
#> [11] "public void org.apache.poi.xssf.usermodel.XSSFRow.setHeight(short)"                                                                                
#> [12] "public short org.apache.poi.xssf.usermodel.XSSFRow.getFirstCellNum()"                                                                              
#> [13] "public org.openxmlformats.schemas.spreadsheetml.x2006.main.CTRow org.apache.poi.xssf.usermodel.XSSFRow.getCTRow()"                                 
#> [14] "public void org.apache.poi.xssf.usermodel.XSSFRow.setRowNum(int)"                                                                                  
#> [15] "public org.apache.poi.xssf.usermodel.XSSFCell org.apache.poi.xssf.usermodel.XSSFRow.createCell(int)"                                               
#> [16] "public org.apache.poi.ss.usermodel.Cell org.apache.poi.xssf.usermodel.XSSFRow.createCell(int)"                                                     
#> [17] "public org.apache.poi.ss.usermodel.Cell org.apache.poi.xssf.usermodel.XSSFRow.createCell(int,int)"                                                 
#> [18] "public org.apache.poi.xssf.usermodel.XSSFCell org.apache.poi.xssf.usermodel.XSSFRow.createCell(int,int)"                                           
#> [19] "public java.lang.String org.apache.poi.xssf.usermodel.XSSFRow.toString()"                                                                          
#> [20] "public int org.apache.poi.xssf.usermodel.XSSFRow.compareTo(java.lang.Object)"                                                                      
#> [21] "public int org.apache.poi.xssf.usermodel.XSSFRow.compareTo(org.apache.poi.xssf.usermodel.XSSFRow)"                                                 
#> [22] "public java.util.Iterator org.apache.poi.xssf.usermodel.XSSFRow.iterator()"                                                                        
#> [23] "public float org.apache.poi.xssf.usermodel.XSSFRow.getHeightInPoints()"                                                                            
#> [24] "public short org.apache.poi.xssf.usermodel.XSSFRow.getLastCellNum()"                                                                               
#> [25] "public void org.apache.poi.xssf.usermodel.XSSFRow.setZeroHeight(boolean)"                                                                          
#> [26] "public void org.apache.poi.xssf.usermodel.XSSFRow.setHeightInPoints(float)"                                                                        
#> [27] "public boolean org.apache.poi.xssf.usermodel.XSSFRow.getZeroHeight()"                                                                              
#> [28] "public boolean org.apache.poi.xssf.usermodel.XSSFRow.isFormatted()"                                                                                
#> [29] "public short org.apache.poi.xssf.usermodel.XSSFRow.getHeight()"                                                                                    
#> [30] "public int org.apache.poi.xssf.usermodel.XSSFRow.getRowNum()"                                                                                      
#> [31] "public org.apache.poi.xssf.usermodel.XSSFSheet org.apache.poi.xssf.usermodel.XSSFRow.getSheet()"                                                   
#> [32] "public org.apache.poi.ss.usermodel.Sheet org.apache.poi.xssf.usermodel.XSSFRow.getSheet()"                                                         
#> [33] "public final native void java.lang.Object.wait(long) throws java.lang.InterruptedException"                                                        
#> [34] "public final void java.lang.Object.wait(long,int) throws java.lang.InterruptedException"                                                           
#> [35] "public final void java.lang.Object.wait() throws java.lang.InterruptedException"                                                                   
#> [36] "public boolean java.lang.Object.equals(java.lang.Object)"                                                                                          
#> [37] "public native int java.lang.Object.hashCode()"                                                                                                     
#> [38] "public final native java.lang.Class java.lang.Object.getClass()"                                                                                   
#> [39] "public final native void java.lang.Object.notify()"                                                                                                
#> [40] "public final native void java.lang.Object.notifyAll()"                                                                                             
#> [41] "public default java.util.Spliterator java.lang.Iterable.spliterator()"                                                                             
#> [42] "public default void java.lang.Iterable.forEach(java.util.function.Consumer)"                                                                       

# for example
rows[[1]]$getRowNum()   # zero based index in Java
#> [1] 0

removeRow(sheet, rows)  # remove them all

# create some row
rows  <- createRow(sheet, rowIndex=1:5)
setRowHeight( rows, multiplier=3)  # 3 times bigger rows than the default