Functions to manipulate worksheets.

Functions to manipulate worksheets.

getSheets(wb)

createSheet(wb, sheetName = "Sheet1")

removeSheet(wb, sheetName = "Sheet1")

Arguments

wb

a workbook object as returned by createWorksheet or loadWorksheet.

sheetName

a character specifying the name of the worksheet to create, or remove.

Value

createSheet returns the created Sheet object. getSheets returns a list of java object references each pointing to an worksheet. The list is named with the sheet names.

See also

Author

Adrian Dragulescu

Examples

file <- system.file("tests", "test_import.xlsx", package = "xlsx")

wb <- loadWorkbook(file)
sheets <- getSheets(wb)

sheet  <- sheets[[2]]  # extract the second sheet

# see all the available java methods that you can call
rJava::.jmethods(sheet)
#>   [1] "public java.util.List org.apache.poi.xssf.usermodel.XSSFSheet.getTables()"                                                                                          
#>   [2] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setTabColor(int)"                                                                                               
#>   [3] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setForceFormulaRecalculation(boolean)"                                                                          
#>   [4] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.getForceFormulaRecalculation()"                                                                              
#>   [5] "public int org.apache.poi.xssf.usermodel.XSSFSheet.findEndOfRowOutlineGroup(int)"                                                                                   
#>   [6] "public static void org.apache.poi.xssf.usermodel.XSSFSheet.setCellComment(java.lang.String,org.apache.poi.xssf.usermodel.XSSFComment)"                              
#>   [7] "public void org.apache.poi.xssf.usermodel.XSSFSheet.addHyperlink(org.apache.poi.xssf.usermodel.XSSFHyperlink)"                                                      
#>   [8] "public org.apache.poi.xssf.usermodel.XSSFRow org.apache.poi.xssf.usermodel.XSSFSheet.createRow(int)"                                                                
#>   [9] "public org.apache.poi.ss.usermodel.Row org.apache.poi.xssf.usermodel.XSSFSheet.createRow(int)"                                                                      
#>  [10] "public void org.apache.poi.xssf.usermodel.XSSFSheet.removeRow(org.apache.poi.ss.usermodel.Row)"                                                                     
#>  [11] "public org.apache.poi.ss.usermodel.Row org.apache.poi.xssf.usermodel.XSSFSheet.getRow(int)"                                                                         
#>  [12] "public org.apache.poi.xssf.usermodel.XSSFRow org.apache.poi.xssf.usermodel.XSSFSheet.getRow(int)"                                                                   
#>  [13] "public int org.apache.poi.xssf.usermodel.XSSFSheet.getPhysicalNumberOfRows()"                                                                                       
#>  [14] "public int org.apache.poi.xssf.usermodel.XSSFSheet.getFirstRowNum()"                                                                                                
#>  [15] "public int org.apache.poi.xssf.usermodel.XSSFSheet.getLastRowNum()"                                                                                                 
#>  [16] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setColumnHidden(int,boolean)"                                                                                   
#>  [17] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isColumnHidden(int)"                                                                                         
#>  [18] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setRightToLeft(boolean)"                                                                                        
#>  [19] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isRightToLeft()"                                                                                             
#>  [20] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setColumnWidth(int,int)"                                                                                        
#>  [21] "public int org.apache.poi.xssf.usermodel.XSSFSheet.getColumnWidth(int)"                                                                                             
#>  [22] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setDefaultColumnWidth(int)"                                                                                     
#>  [23] "public int org.apache.poi.xssf.usermodel.XSSFSheet.getDefaultColumnWidth()"                                                                                         
#>  [24] "public short org.apache.poi.xssf.usermodel.XSSFSheet.getDefaultRowHeight()"                                                                                         
#>  [25] "public float org.apache.poi.xssf.usermodel.XSSFSheet.getDefaultRowHeightInPoints()"                                                                                 
#>  [26] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setDefaultRowHeight(short)"                                                                                     
#>  [27] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setDefaultRowHeightInPoints(float)"                                                                             
#>  [28] "public org.apache.poi.ss.usermodel.CellStyle org.apache.poi.xssf.usermodel.XSSFSheet.getColumnStyle(int)"                                                           
#>  [29] "public int org.apache.poi.xssf.usermodel.XSSFSheet.addMergedRegion(org.apache.poi.ss.util.CellRangeAddress)"                                                        
#>  [30] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setVerticallyCenter(boolean)"                                                                                   
#>  [31] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setHorizontallyCenter(boolean)"                                                                                 
#>  [32] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.getHorizontallyCenter()"                                                                                     
#>  [33] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.getVerticallyCenter()"                                                                                       
#>  [34] "public void org.apache.poi.xssf.usermodel.XSSFSheet.removeMergedRegion(int)"                                                                                        
#>  [35] "public int org.apache.poi.xssf.usermodel.XSSFSheet.getNumMergedRegions()"                                                                                           
#>  [36] "public org.apache.poi.ss.util.CellRangeAddress org.apache.poi.xssf.usermodel.XSSFSheet.getMergedRegion(int)"                                                        
#>  [37] "public java.util.Iterator org.apache.poi.xssf.usermodel.XSSFSheet.rowIterator()"                                                                                    
#>  [38] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setAutobreaks(boolean)"                                                                                         
#>  [39] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setDisplayGuts(boolean)"                                                                                        
#>  [40] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setDisplayZeros(boolean)"                                                                                       
#>  [41] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isDisplayZeros()"                                                                                            
#>  [42] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setFitToPage(boolean)"                                                                                          
#>  [43] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setRowSumsBelow(boolean)"                                                                                       
#>  [44] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setRowSumsRight(boolean)"                                                                                       
#>  [45] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.getAutobreaks()"                                                                                             
#>  [46] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.getDisplayGuts()"                                                                                            
#>  [47] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.getFitToPage()"                                                                                              
#>  [48] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.getRowSumsBelow()"                                                                                           
#>  [49] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.getRowSumsRight()"                                                                                           
#>  [50] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isPrintGridlines()"                                                                                          
#>  [51] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setPrintGridlines(boolean)"                                                                                     
#>  [52] "public org.apache.poi.xssf.usermodel.XSSFPrintSetup org.apache.poi.xssf.usermodel.XSSFSheet.getPrintSetup()"                                                        
#>  [53] "public org.apache.poi.ss.usermodel.PrintSetup org.apache.poi.xssf.usermodel.XSSFSheet.getPrintSetup()"                                                              
#>  [54] "public org.apache.poi.ss.usermodel.Header org.apache.poi.xssf.usermodel.XSSFSheet.getHeader()"                                                                      
#>  [55] "public org.apache.poi.ss.usermodel.Footer org.apache.poi.xssf.usermodel.XSSFSheet.getFooter()"                                                                      
#>  [56] "public double org.apache.poi.xssf.usermodel.XSSFSheet.getMargin(short)"                                                                                             
#>  [57] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setMargin(short,double)"                                                                                        
#>  [58] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.getProtect()"                                                                                                
#>  [59] "public void org.apache.poi.xssf.usermodel.XSSFSheet.protectSheet(java.lang.String)"                                                                                 
#>  [60] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.getScenarioProtect()"                                                                                        
#>  [61] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setZoom(int)"                                                                                                   
#>  [62] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setZoom(int,int)"                                                                                               
#>  [63] "public short org.apache.poi.xssf.usermodel.XSSFSheet.getTopRow()"                                                                                                   
#>  [64] "public short org.apache.poi.xssf.usermodel.XSSFSheet.getLeftCol()"                                                                                                  
#>  [65] "public void org.apache.poi.xssf.usermodel.XSSFSheet.showInPane(short,short)"                                                                                        
#>  [66] "public void org.apache.poi.xssf.usermodel.XSSFSheet.showInPane(int,int)"                                                                                            
#>  [67] "public void org.apache.poi.xssf.usermodel.XSSFSheet.shiftRows(int,int,int)"                                                                                         
#>  [68] "public void org.apache.poi.xssf.usermodel.XSSFSheet.shiftRows(int,int,int,boolean,boolean)"                                                                         
#>  [69] "public void org.apache.poi.xssf.usermodel.XSSFSheet.createFreezePane(int,int,int,int)"                                                                              
#>  [70] "public void org.apache.poi.xssf.usermodel.XSSFSheet.createFreezePane(int,int)"                                                                                      
#>  [71] "public void org.apache.poi.xssf.usermodel.XSSFSheet.createSplitPane(int,int,int,int,int)"                                                                           
#>  [72] "public org.apache.poi.hssf.util.PaneInformation org.apache.poi.xssf.usermodel.XSSFSheet.getPaneInformation()"                                                       
#>  [73] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setDisplayGridlines(boolean)"                                                                                   
#>  [74] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isDisplayGridlines()"                                                                                        
#>  [75] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setDisplayFormulas(boolean)"                                                                                    
#>  [76] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isDisplayFormulas()"                                                                                         
#>  [77] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setDisplayRowColHeadings(boolean)"                                                                              
#>  [78] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isDisplayRowColHeadings()"                                                                                   
#>  [79] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setRowBreak(int)"                                                                                               
#>  [80] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isRowBroken(int)"                                                                                            
#>  [81] "public void org.apache.poi.xssf.usermodel.XSSFSheet.removeRowBreak(int)"                                                                                            
#>  [82] "public int[] org.apache.poi.xssf.usermodel.XSSFSheet.getRowBreaks()"                                                                                                
#>  [83] "public int[] org.apache.poi.xssf.usermodel.XSSFSheet.getColumnBreaks()"                                                                                             
#>  [84] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setColumnBreak(int)"                                                                                            
#>  [85] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isColumnBroken(int)"                                                                                         
#>  [86] "public void org.apache.poi.xssf.usermodel.XSSFSheet.removeColumnBreak(int)"                                                                                         
#>  [87] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setColumnGroupCollapsed(int,boolean)"                                                                           
#>  [88] "public java.lang.String org.apache.poi.xssf.usermodel.XSSFSheet.getSheetName()"                                                                                     
#>  [89] "public org.openxmlformats.schemas.spreadsheetml.x2006.main.CTWorksheet org.apache.poi.xssf.usermodel.XSSFSheet.getCTWorksheet()"                                    
#>  [90] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setSelected(boolean)"                                                                                           
#>  [91] "public java.util.Iterator org.apache.poi.xssf.usermodel.XSSFSheet.iterator()"                                                                                       
#>  [92] "public org.apache.poi.ss.usermodel.Footer org.apache.poi.xssf.usermodel.XSSFSheet.getOddFooter()"                                                                   
#>  [93] "public org.apache.poi.ss.usermodel.Footer org.apache.poi.xssf.usermodel.XSSFSheet.getEvenFooter()"                                                                  
#>  [94] "public org.apache.poi.ss.usermodel.Footer org.apache.poi.xssf.usermodel.XSSFSheet.getFirstFooter()"                                                                 
#>  [95] "public org.apache.poi.ss.usermodel.Header org.apache.poi.xssf.usermodel.XSSFSheet.getOddHeader()"                                                                   
#>  [96] "public org.apache.poi.ss.usermodel.Header org.apache.poi.xssf.usermodel.XSSFSheet.getEvenHeader()"                                                                  
#>  [97] "public org.apache.poi.ss.usermodel.Header org.apache.poi.xssf.usermodel.XSSFSheet.getFirstHeader()"                                                                 
#>  [98] "public int org.apache.poi.xssf.usermodel.XSSFSheet.getNumHyperlinks()"                                                                                              
#>  [99] "public org.apache.poi.xssf.usermodel.helpers.ColumnHelper org.apache.poi.xssf.usermodel.XSSFSheet.getColumnHelper()"                                                
#> [100] "public org.apache.poi.xssf.usermodel.XSSFComment org.apache.poi.xssf.usermodel.XSSFSheet.createComment()"                                                           
#> [101] "public org.apache.poi.xssf.usermodel.XSSFHyperlink org.apache.poi.xssf.usermodel.XSSFSheet.getHyperlink(int,int)"                                                   
#> [102] "public org.apache.poi.ss.usermodel.Drawing org.apache.poi.xssf.usermodel.XSSFSheet.createDrawingPatriarch()"                                                        
#> [103] "public org.apache.poi.xssf.usermodel.XSSFDrawing org.apache.poi.xssf.usermodel.XSSFSheet.createDrawingPatriarch()"                                                  
#> [104] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setRepeatingRows(org.apache.poi.ss.util.CellRangeAddress)"                                                      
#> [105] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setRepeatingColumns(org.apache.poi.ss.util.CellRangeAddress)"                                                   
#> [106] "public java.lang.String org.apache.poi.xssf.usermodel.XSSFSheet.getActiveCell()"                                                                                    
#> [107] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setActiveCell(java.lang.String)"                                                                                
#> [108] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.hasComments()"                                                                                               
#> [109] "public org.openxmlformats.schemas.spreadsheetml.x2006.main.CTCellFormula org.apache.poi.xssf.usermodel.XSSFSheet.getSharedFormula(int)"                             
#> [110] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isAutoFilterLocked()"                                                                                        
#> [111] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isDeleteColumnsLocked()"                                                                                     
#> [112] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isDeleteRowsLocked()"                                                                                        
#> [113] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isFormatCellsLocked()"                                                                                       
#> [114] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isFormatColumnsLocked()"                                                                                     
#> [115] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isFormatRowsLocked()"                                                                                        
#> [116] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isInsertColumnsLocked()"                                                                                     
#> [117] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isInsertHyperlinksLocked()"                                                                                  
#> [118] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isInsertRowsLocked()"                                                                                        
#> [119] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isPivotTablesLocked()"                                                                                       
#> [120] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isSortLocked()"                                                                                              
#> [121] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isObjectsLocked()"                                                                                           
#> [122] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isScenariosLocked()"                                                                                         
#> [123] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isSelectLockedCellsLocked()"                                                                                 
#> [124] "public org.apache.poi.ss.usermodel.Workbook org.apache.poi.xssf.usermodel.XSSFSheet.getWorkbook()"                                                                  
#> [125] "public org.apache.poi.xssf.usermodel.XSSFWorkbook org.apache.poi.xssf.usermodel.XSSFSheet.getWorkbook()"                                                            
#> [126] "public void org.apache.poi.xssf.usermodel.XSSFSheet.groupColumn(int,int)"                                                                                           
#> [127] "public void org.apache.poi.xssf.usermodel.XSSFSheet.ungroupColumn(int,int)"                                                                                         
#> [128] "public void org.apache.poi.xssf.usermodel.XSSFSheet.groupRow(int,int)"                                                                                              
#> [129] "public void org.apache.poi.xssf.usermodel.XSSFSheet.ungroupRow(int,int)"                                                                                            
#> [130] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setRowGroupCollapsed(int,boolean)"                                                                              
#> [131] "public void org.apache.poi.xssf.usermodel.XSSFSheet.setDefaultColumnStyle(int,org.apache.poi.ss.usermodel.CellStyle)"                                               
#> [132] "public void org.apache.poi.xssf.usermodel.XSSFSheet.autoSizeColumn(int)"                                                                                            
#> [133] "public void org.apache.poi.xssf.usermodel.XSSFSheet.autoSizeColumn(int,boolean)"                                                                                    
#> [134] "public org.apache.poi.xssf.usermodel.XSSFComment org.apache.poi.xssf.usermodel.XSSFSheet.getCellComment(int,int)"                                                   
#> [135] "public org.apache.poi.ss.usermodel.Comment org.apache.poi.xssf.usermodel.XSSFSheet.getCellComment(int,int)"                                                         
#> [136] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isSelected()"                                                                                                
#> [137] "public org.apache.poi.ss.usermodel.CellRange org.apache.poi.xssf.usermodel.XSSFSheet.setArrayFormula(java.lang.String,org.apache.poi.ss.util.CellRangeAddress)"     
#> [138] "public org.apache.poi.ss.usermodel.CellRange org.apache.poi.xssf.usermodel.XSSFSheet.removeArrayFormula(org.apache.poi.ss.usermodel.Cell)"                          
#> [139] "public org.apache.poi.ss.usermodel.DataValidationHelper org.apache.poi.xssf.usermodel.XSSFSheet.getDataValidationHelper()"                                          
#> [140] "public void org.apache.poi.xssf.usermodel.XSSFSheet.addValidationData(org.apache.poi.ss.usermodel.DataValidation)"                                                  
#> [141] "public org.apache.poi.xssf.usermodel.XSSFAutoFilter org.apache.poi.xssf.usermodel.XSSFSheet.setAutoFilter(org.apache.poi.ss.util.CellRangeAddress)"                 
#> [142] "public org.apache.poi.ss.usermodel.AutoFilter org.apache.poi.xssf.usermodel.XSSFSheet.setAutoFilter(org.apache.poi.ss.util.CellRangeAddress)"                       
#> [143] "public org.apache.poi.xssf.usermodel.XSSFSheetConditionalFormatting org.apache.poi.xssf.usermodel.XSSFSheet.getSheetConditionalFormatting()"                        
#> [144] "public org.apache.poi.ss.usermodel.SheetConditionalFormatting org.apache.poi.xssf.usermodel.XSSFSheet.getSheetConditionalFormatting()"                              
#> [145] "public org.apache.poi.ss.util.CellRangeAddress org.apache.poi.xssf.usermodel.XSSFSheet.getRepeatingRows()"                                                          
#> [146] "public org.apache.poi.ss.util.CellRangeAddress org.apache.poi.xssf.usermodel.XSSFSheet.getRepeatingColumns()"                                                       
#> [147] "public java.util.List org.apache.poi.xssf.usermodel.XSSFSheet.getDataValidations()"                                                                                 
#> [148] "public org.apache.poi.xssf.usermodel.XSSFTable org.apache.poi.xssf.usermodel.XSSFSheet.createTable()"                                                               
#> [149] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isSelectUnlockedCellsLocked()"                                                                               
#> [150] "public boolean org.apache.poi.xssf.usermodel.XSSFSheet.isSheetLocked()"                                                                                             
#> [151] "public void org.apache.poi.xssf.usermodel.XSSFSheet.enableLocking()"                                                                                                
#> [152] "public void org.apache.poi.xssf.usermodel.XSSFSheet.disableLocking()"                                                                                               
#> [153] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockAutoFilter()"                                                                                               
#> [154] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockDeleteColumns()"                                                                                            
#> [155] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockDeleteRows()"                                                                                               
#> [156] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockFormatCells()"                                                                                              
#> [157] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockFormatColumns()"                                                                                            
#> [158] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockFormatRows()"                                                                                               
#> [159] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockInsertColumns()"                                                                                            
#> [160] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockInsertHyperlinks()"                                                                                         
#> [161] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockInsertRows()"                                                                                               
#> [162] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockPivotTables()"                                                                                              
#> [163] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockSort()"                                                                                                     
#> [164] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockObjects()"                                                                                                  
#> [165] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockScenarios()"                                                                                                
#> [166] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockSelectLockedCells()"                                                                                        
#> [167] "public void org.apache.poi.xssf.usermodel.XSSFSheet.lockSelectUnlockedCells()"                                                                                      
#> [168] "public java.lang.String org.apache.poi.POIXMLDocumentPart.toString()"                                                                                               
#> [169] "public final org.apache.poi.POIXMLDocumentPart org.apache.poi.POIXMLDocumentPart.getParent()"                                                                       
#> [170] "public final void org.apache.poi.POIXMLDocumentPart.addRelation(java.lang.String,org.apache.poi.POIXMLDocumentPart)"                                                
#> [171] "public final org.apache.poi.POIXMLDocumentPart org.apache.poi.POIXMLDocumentPart.getRelationById(java.lang.String)"                                                 
#> [172] "public final java.lang.String org.apache.poi.POIXMLDocumentPart.getRelationId(org.apache.poi.POIXMLDocumentPart)"                                                   
#> [173] "public final org.apache.poi.openxml4j.opc.PackagePart org.apache.poi.POIXMLDocumentPart.getPackagePart()"                                                           
#> [174] "public final java.util.List org.apache.poi.POIXMLDocumentPart.getRelations()"                                                                                       
#> [175] "public final org.apache.poi.openxml4j.opc.PackageRelationship org.apache.poi.POIXMLDocumentPart.getPackageRelationship()"                                           
#> [176] "public final org.apache.poi.POIXMLDocumentPart org.apache.poi.POIXMLDocumentPart.createRelationship(org.apache.poi.POIXMLRelation,org.apache.poi.POIXMLFactory)"    
#> [177] "public final org.apache.poi.POIXMLDocumentPart org.apache.poi.POIXMLDocumentPart.createRelationship(org.apache.poi.POIXMLRelation,org.apache.poi.POIXMLFactory,int)"
#> [178] "public final native void java.lang.Object.wait(long) throws java.lang.InterruptedException"                                                                         
#> [179] "public final void java.lang.Object.wait(long,int) throws java.lang.InterruptedException"                                                                            
#> [180] "public final void java.lang.Object.wait() throws java.lang.InterruptedException"                                                                                    
#> [181] "public boolean java.lang.Object.equals(java.lang.Object)"                                                                                                           
#> [182] "public native int java.lang.Object.hashCode()"                                                                                                                      
#> [183] "public final native java.lang.Class java.lang.Object.getClass()"                                                                                                    
#> [184] "public final native void java.lang.Object.notify()"                                                                                                                 
#> [185] "public final native void java.lang.Object.notifyAll()"                                                                                                              
#> [186] "public default java.util.Spliterator java.lang.Iterable.spliterator()"                                                                                              
#> [187] "public default void java.lang.Iterable.forEach(java.util.function.Consumer)"                                                                                        

# for example
sheet$getLastRowNum()
#> [1] 12