Read a contiguous set of rows into an R matrix

Read a contiguous set of rows into an R character matrix. Uses the RInterface for speed.

readRows(sheet, startRow, endRow, startColumn, endColumn = NULL)

Arguments

sheet

a Worksheet object.

startRow

a numeric value for the starting row.

endRow

a numeric value for the ending row. If NULL it reads all the rows in the sheet.

startColumn

a numeric value for the starting column.

endColumn

a numeric value for the ending column. Empty cells will be returned as "".

Value

A character matrix.

Details

Use the readRows function when you want to read a row or a block block of data from an Excel worksheet. Internally, the loop over rows is done in R, and the loop over columns is done in Java, so this function achieves good performance when number of rows << number of columns.

In general, you should prefer the function readColumns over this one.

See also

Author

Adrian Dragulescu

Examples

if (FALSE) {

  file <- system.file("tests", "test_import.xlsx", package = "xlsx")

  wb     <- loadWorkbook(file)
  sheets <- getSheets(wb)

  sheet <- sheets[["all"]]
  res <- readRows(sheet, startRow=3, endRow=7, startColumn=3, endColumn=10)

}